requirements.txt
google-generativeai
customtkinter
pyinstaller
python-dotenv
